Move more manifest functionality into ManifestUtil
